(GH-723) Fully qualify MsiExec
In commit cfa8e2fd, we introduced  introduced a check in
`Start-ChocolateyProcessAsAdmin`  that determines if an executable
exists. Unfortunately this function doesn't try to verify if the item
exists at a location on the PATH, thus breaking calls from
`Install-ChocolateyInstallPackage` and `Uninstall-ChocolateyPackage`
that call this function with simply 'msiexec'.

We should fully qualify msiexec as a security measure, but we should
also not fail on missing if msiexec is the executable we are trying
to run.