(GH-739) Get-UninstallRegistryKey / update templates
Add Get-UninstallRegistryKey function to provide a means of getting
uninstaller registry keys that does not fail with incorrectly
encoded keys and values.

Update the code documentation to link update the function with other
similar functions and generate updated documentation

Add the function to the uninstall template and nuspec with
instructions to use the chocolatey-uninstall.extension package when
supporting 0.9.9.x and below as the new function will only be
available in 0.9.10 and above.

The helper includes a trycatch for invalid registry keys. Using the
catch is a requirement as blindly enumerating will fail every time
if even one invalid key exists.