(GH-741) Ensure CultureInfo with PowerShell Host
When using the built-in posh host, the culture information that is
returned should be the user's culture info, so that gathering the LCID
for language settings in packages works appropriately. Without this
fix, everyone will be given the invariant culture (127), which could
cause problems across every package that detects locale.