(GH-733) Don't pass some args to dependencies
When using Install arguments or package parameters, those only apply to
the specified packages and not to the dependent packages, unless those
packages are the meta/virtual packages that come in "*.install",
"*.portable" flavors. If a package doesn't meet the criteria, don't
pass install arguments or package parameters to that package.