(GH-622) Remove NuGet temp folders
`TEMP\NuGetScratch` and sometimes a `TEMP\x\Nuget` folder are created
by NuGet Core, however it doesn't clean up after itself on this. It is
expected for you to call `OptimizedZipPackage.PurgeCache()`, but this
does not appear to actually remove the folder because PurgeCache is
dependent on the GUID that it creates for the subdirectory. It doesn't
hold a reference to that and so it comes up with one that doesn't exist
by the time we call it.

Handle cleaning up those caches at the end of most runs.