(GH-751) Use package name/version from env vars
Previously, there was a small issue that choco would not be able to
force a package to re-download if it could not find the files in the
local cache. This is due to using the user passed parameter for package
name instead of the actual package name where the code is looking for
the package. This is bad because it depends on the user to pass the
correct parameter and that may not be the case.

Instead of trusting input, make the pathing deterministic by using the
environment variables that choco set up for those values instead.