(GH-732) Request/Response Timeout configurable
Request timeout was set at around 20 seconds, which doesn't help for
slow internet access when running choco. Make the value configurable
and set a default of 45 seconds in the configuration.

Additionally, the response timeout (request.ReadWriteTimeout) was set
at the default of 5 minutes since it was not specified. This means that
if something took longer than 5 minutes to download, it timed out.
Instead of doing that, we should give it the full time of command
execution timeout to finish the response. If it does use that full
time, the package will fail, but the user could then bump the execution
timeout on the next call. It's a generous 45 minutes by default, so it
should reduce/remove any issues that may have been occurring.