(GH-753) Get-WebFileName Matches on Invalid Chars
Previously, the regex used to determine whether invalid characters were
used or not used the option `IgnorePatternWhitespace` and included
spaces in the way that it returned the values. This actually caused it
to match on almost nothing.

Instead, remove the whitespace when joining
`Path.GetInvalidFileNameChars` and additionally add `=;`. While equals
and semi-colon are valid for file name, they are indicative of an issue
with the file name.

At every attempt to determine the file name, run the regex to determine
if what was determined was good to go or if the next check needs to
run. This allows as many of the checks to occur as necessary to
determine the file name from the remote location.

If any method of attempting to get the file name fails, fall back to
the default filename passed into the method.