(GH-198) Set and remove a pending file
Once a package install/upgrade is passed from NuGet Core to Chocolatey,
we are in a pending state where the package files exist but more work
is to be done. Create a pending file for the package that tells
Chocolatey that the package is not yet finished installing in case it
gets interrupted for some reason. Also ensure that the pending file is
cleaned up if the package installation is successful.