(GH-708) PowerShell Host.Version Returns Actual
Previously, when calling `$host.version` using the built-in posh host,
the current version of Chocolatey would be returned. Depending on
`$host.version` is not a recommended best practice for PowerShell,
folks should depend on `$PSVersionTable` instead. However, not everyone
knows about best practices, so in practice folks are likely using
`$host.version` because it is easy to remember.

Due to the fact that some folks are using `$host.version` as a way of
building a compatibility check, ensure that choco returns the actual
version of PowerShell that is installed.