(GH-735) Remove actual when uninstalling meta
When uninstalling a meta package, a user may not realize they need to
uninstall the actual package as well, such as a *.portable or *.intall.
For instance, if a user uninstalls git package, they may not realize
they have only removed the meta package and git.install and
subsequently, Git for Windows are still installed. Provide a warning
with a timeout offering to add the actual package to the list. If there
is no response or the answer is no, add the actual package to the list
of warnings for the package summary with an explanation on how to
uninstall that package.