(GH-769) API - Resolve Merged Assemblies
When using the API, it should resolve merged assemblies but it may
not if there is an assembly handler registered. Before loading up
code that resolves external assemblies, register an AssemblyResolve
handler with the purpose of determining if the assembly has
Chocolatey's Public Key token or not. If it does, use Chocolatey's
merged assembly instead of the conflicting assembly.