(GH-446) Opt out checksum tool from FIPS policy
Opt out the checksum.exe tool from FIPS algorithm policy because it is
not in violation of the standard - because it doesn't do anything with
encryption or crypto, other than hash a file to determine a way to
uniquely identify it. Unfortunately the .NET Framework doesn't have a
good way to verify this is the case and the Managed cryptos
(SHA256Managed vs SHA256Cng) were never submitted to NIST, so they are
not considered compliant. That's somewhat beside the point - it
doesn't matter what is used for checksumming, it doesn't violate FIPS
because it is only checksumming.

"The .NET Framework's enforcement of FIPS mode cannot tell whether any
particular use of a cryptographic class is not for security purposes
and thus not in violation of standards." - Aaron Margosis, MSFT, "Why
We're Not Recommending 'FIPS Mode' Anymore" -
https://blogs.technet.microsoft.com/secguide/2014/04/07/why-were-not-recommending-fips-mode-anymore/