(GH-439) Refresh Environment Before Package Scripts
Before running each package script, refresh the process environment
variables so that the subprocess will inherit those. This allows
updates to PATH and additional environment variables to be taken
advantage of in packages that may depend on them being set by a package
dependency that installs during the same run.