(GH-759) Summarize affected pkgs when at least 5
When installing from a packages.config, or upgrading/uninstalling all
packages, things can get lost in the information about things that were
successful. It's better to report this in a summary when there are more
than an easily scannable number of packages involved. When the number
is 5 or more, report the summary of affected packages.