(GH-732) reduce default webrequest timeout to 30s
This time seems more appropriate. Also, sometimes there is a bit of a
hang when calling for a response. It seems if you don't attempt to set
the fully qualified name of `$res` and just assign the variable, it
will error early instead of hanging. This could just be a local issue.