(GH-822) Provide pending override and file wait
When a pending file is created, it needs to be at least 10 seconds old
before choco will automatically remove the unfinished pacakge. This
allows running choco functions inside of a package (as long as they are
done quickly).

If someone needs to override the functionality of the removal of a
package due to the pending file, choco will also look for a
.chocolateyPendingSkip and skip the removal if that file is found.