(GH-826) SYSTEM user always uses machine TEMP
Using `Path.GetTempPath()` instead of
`Environment.GetEnvironmentVariable("TEMP")` in 76ae7e2f7 for GH-532
has caused the SYSTEM user to return one of the following:

 * `c:\windows\system32\config\systemprofile\appdata\local\temp`
 * `c:\windows\syswow64\config\systemprofile\appdata\local\temp`

These folders seem to cause issues with accessing files, causing all
kinds of fun issues, such as "Not able to read package from path" and
not able to find executables being downloaded to these folders.

If one of these two folders is detected, ensure that the folder
structure is `%SystemRoot%\TEMP`.