(GH-857) Fix - Sometimes paths contain null chars
When values coming from the registry are not cleansed prior to being
passed to Start-ChocolateyProcessAsAdmin, they sometimes can contain
null characters. Remove those characters in the helper.

Further, if attempting to get the full path for looking to see if the
file is a text file instead of an executable, do not fail the run.