(GH-872) Fix: Download progress never reaches completion
Always pass -PercentComplete to Write-Progress, otherwise the output will
not be logged due to the -1 check. Because we can reach 100% when there
are a few bytes left to retrieve, allow further logging of progress even if
we reached 100%.

In Get-WebFile and Get-FtpFile, don't log completion until we read 0 bytes
from the stream. Also, when calculating the percentage, round down instead
of up so we don't log 100% until we are actually at 100%.