(GH-883) Wait Up to 15 Seconds for Process Exit
Even though `$process.WaitForExit()` has identified that the process
has exited, it appears that the process may be performing some final
cleanup as the exit code is still null when it is passed to
`Set-PowerShellExitCode`. Allow the process to take up to 15 seconds to
fully exit so that it has cleaned up. In the case of 7zip, ensure that
it has ample time to finish extracting after it has indicated that it
has completed the extraction and is exiting.