(GH-112) Require checksums
Checksums are now required when downloading remote resources. This
provides safety when using Chocolatey with packages that may not be
safe. Any remote resource requires a checksum, as both HTTP and HTTPS
sources could be compromised and the downloaded binaries need to be
validated for integrity against what was originally intended by the
packager.

A user can choose to provide their own or override the default safety
mechanism with `--allow-empty-checksums`, but this is a user decision
and not solely a package maintainer decision anymore.