(GH-839) Switch to apply package parameters to dependent packages
In version 0.9.9.12 and earlier, install arguments and package-parameters would also be applied to dependent packages, but in choco 0.9.10 this was changed so that it would only be applied to the specified package.
Sometimes it will be useful to pass the install arguments and package parameters further to dependent packages. This can be controlled by using cmdline switches.