(GH-458) Support TLS v1.2
The .NET Framework supports TLS v1.2 as of version 4.5 -
http://msdn.microsoft.com/en-us/library/system.security.authentication.sslprotocols(v=vs.110).aspx
but the security protocol is not set to use it out of the box.

Set the SecurityProtocol to start with the highest encryption available
and move down from there.

Choco is compiled on .NET 4.0, but .NET 4.5 is an in place upgrade,
which gives us access to set the proper security protocols by
converting the enumeration values directly into `SecurityProtocolType`.
This will fail when running on .NET 4.0, so fall back to using TLS,
then SSLv3 with a warning recommending folks upgrade to .NET 4.5 at
their earliest convenience.