(docs) move GenerateDocs.ps1 / update
Move GenerateDocs.ps1 to top level for ease of use. Update
the directory so that it is looking from top level instead of
needing to back up a directory first to find the top level
and ensure pro override doesn't override the word "provides"
or other words that could be accidentally partially
overridden.