(GH-960) Fix: 7z File System Redirection
When using a 32-bit version of 7z.exe to install applications, it can
be subject to file system redirection. This was originally handled
since 0.9.8.24 with
https://github.com/chocolatey/chocolatey/commit/54985d5069467dfaea037809568ec1a1b9acdd11
but was later accidentally reverted in choco v0.9.10 - GH-616
(bde0941d7).

With redirection, calls in 32bit processes to C:\Windows\System32 are
automatically redirected to C:\Windows\SysWOW64. This can have
undesirable effects when the path is meant to look for System32.
Windows provides `SysNative` for this reason, and it should be used if
the path is found (32bit processes on 64bit Windows). This puts the old
behavior back to how it worked prior to the changes in 0.9.10.