(doc) provide more awareness about licensed
Add new messages to display at random. A run has a 1 in 10 chance of
showing a message. With 8 different messages, it could take up to 80
runs to display them all if (big if) the messages always came out
unique. More than likely it would take more like 200 runs to see all of
the different messaging.