(GH-934) Support Paths > 260 Characters
Installing a package or working with packaging could mean paths that
are over 260 characters (MAX_PATH). Sometimes backing up a path will
take it over 260 characters resulting in errors when attempting to
upgrade/uninstall packages that have pretty long paths.

This is done by using the AlphaFS library. Unfortunately it doesn't use
the .NET Framework methods and check lengths before it uses its
implementation with native calls so it doesn't appear to be Mono safe.
This meant the implementation of IFileSystem that is being used will
need to perform those checks and fall back on errors.

This also means the addition of dynamics to support switching between
two interpretations of FileInfo and DirectoryInfo.