(GH-945) Fix - Fails Installs Running Without User
When running from a scheduled task or from being run when no user is
logged into the system, choco was failing on a path combine error when
it should have just moved on. This is happening in cleaning up the
nuget cache that was introduced in 0.9.10. The path combine was
attempting to combine `$env:LocalAppData` with other components.
LocalAppData is a user dependent environment variable and empty when no
user is logged in. Instead look at the path to see if it is null and
return early instead of attempting to clean up the cache.