(GH-458) Warn To Verbose Log For Now
We want to provide the TLS warning in a way that doesn't have people
just blindly ignoring future warnings, so we need a way to toggle the
warning off once someone has accepted they have read and understood the
warning. For now we will log to verbose so that it is at least present
in logs, and when folks run at verbose.