(GH-983) Do not add unparsed options to package names
When an unparsed option starts with `-` or `--`, that was not meant to
be a package name so do not add it to the list of packages to install.
A package also cannot start with a dash as its id so it is fine not to
add them to the list.