(GH-992) allow uninstall scripts for side by side
When there are multiple side by side installs, choco should support
automatically determine if it is a side by side install that we are
uninstalling instead of expecting `--sxs` to be passed to the uninstall
commands. Prior to this it only executed the uninstall behavior for the
latest version of the package.