(GH-996) Fix: RefreshEnv command throws errors
With the new call to Write-FunctionCallLogMessage.ps1, the profile was
not updated to use that call. So it fails with lots of red text on the
screen as you attempt to run things like `refreshenv` from PowerShell.
Add the function to be loaded so there are no issues running
`refreshenv`.