(GH-995) Ensure beforeModify runs as first action
For upgrades and uninstalls, there was an attempt to run other
operations ahead of the beforeModify script, which included removing a
rollback directory that may have been locking files for a service.
Instead ensure that before modify is the first physcial operation
attempted prior to upgrades and uninstall actions.

For uninstalls, ensure that the before modify action is passed through
as part of the actions to take prior to uninstall instead of waiting
for the regular means to come back to handle_package_uninstall.