(GH-1000) Don't check $LASTEXITCODE by default
Starting with 33b628525c, which was released as part of GH-512 in
0.9.10, we started checking `$LASTEXITCODE` in addition to the script
command success. This meant it offered the ability to capture when a
script exited with `exit 1` and handle that accordingly. However that
was not a recommended scenario for returning errors from scripts.

Checking `$LastExitCode` checks the last executable's exit code when the
script specifically does not call `Exit`. This can lead to very
perplexing failures, such as running a successful xcopy that exits with
2 and seeing package failures without understanding why. Since it is
not typically recommended to call `Exit` to return a value from
PowerShell because of issues with different hosts, it's less of a
concern to only look at explicit failures.