(GH-676) Fix: Unzip specific folder feature is broken
Previously, when attempting to unzip a subdirectory in a zipped
archive, the Chocolatey unzip functions were ignoring the `SpecificFolder`
parameter. This was caused by a change to use 7za.exe in August 2013 in commit
https://github.com/chocolatey/chocolatey/commit/2e83d2e0cf07bb95b2a73a7a124257b57f009170.

That changed carried over into the newer codebase. More recently there was a
switch back to 7z.exe (full 7zip), so it is possible to unzip a subdirectory
in an archive again. Ensure that `SpecficFolder` is added to the 7zip options
when it is specified.