(GH-1043) Fix: Set-EnvironmentVariable writes error when SYSTEM
Set-EnvironmentVariable uses try/catch to handle when a registry key doesn't
exist. When running as SYSTEM, a local key may not exist, so an error is logged
as a warning.

Add explicit check to see if the environment variable registry
key exists before tring to get its type to avoid logging an error/warning.