(GH-1069) Fix - Disable failOnInvalidOrMissingLicense Feature
Once enabled, if one does not have a license, they cannot run choco feature
to disable the feature and cannot use Chocolatey until one manually changes
the feature in the configuration file.

This change disables the check for a license when
`failOnInvalidOrMissingLicense` is enabled and the 'features' command is
being used.