(GH-1047) Robust/efficient xml serialization
Previously, the xml file serialization could cause corruption when it
did not properly replace the file or when multiple choco processes were
running at the same time. Additionally, there was no backup of the
original config, so the configuration would need to be reset back to
the default configuration and then readjusted.

Switch to evaluating differences in memory instead of an update file.
If the upates are different, then write out the update file and use
replace file to have it create a backup of the old file and replace the
config with the updated version.

When deserializing, if the file is corrupt, look to use the backup and
put it back in place if it works. Handle cases where there is no original
config file and the backup file doesn't exist.

Another benefit of comparing in memory is that no errors are logged for
non-admins when the config update file can't be written.