(GH-1098) Get-FtpFile - change int to long / improve error handling
Previously, using integers in Get-FtpFile could result in integer stack
overflows. Change to long to bring in line with `Get-WebFile` and avoid
errors.

Removed unnecessary catch block, shifted close of reader and writer to
the enclosing finally block. The close calls are wrapped in an empty
try/catch as an error here would mask an error thrown from the catch
block.