(GH-1041) Fix: Skip loading synced dlls as extensions
When Chocolatey loads the extensions folders, sometimes it is getting
caught by the synced folder that may contain DLLs. In newer licensed
versions, this synced folder is kept in a better location, but those
older versions need a way of ensuring that those DLLs are not attempted
to be loaded up.