(GH-906) Fix source path when installing "*.nuspec"
Previously, when the install target was a nuspec file and the output
folder was not set to the folder that contained the nuspec file then
the installation failed because the nupkg file could not be found.
Update sets the sources parameter to the output folder during the
packaging of the nuspec file.