(GH-533) Support packages.config file in upgrade
Implements a subset of the functionality, support for packages.config
files in the upgrade command. Also extended the tests to test the new
code. I did not see any value in implementing package.config support for
uninstall because the problem I am interested in solving is being able
to upgrade multiple packages each with a specific version in one
command. Since uninstall (generally?) doesn't require specifying a
version this can be done on the uninstall command line, i.e. just give
the package list to uninstall.