(GH-1145) Better Logs On Licensed Dll Failures
When there is a failure in loading the licensed assembly or running a
function in the licensed assembly, there should be a way to provide
more detailed information about what happened. Look to see if the
user specified debug as part of the arguments. If so provide the stack
trace. If there is an inner exception, be sure to include that detail
as well.