(GH-262) Sources Explicitly Bypass Proxy
Configure a source to explicitly bypass any configured or system
proxies by specifically providing `bypassProxy` as a parameter for
sources to configure. Without this a source could be subject to a proxy
if one is configured, even if it should not be. This allows a user to
explicitly configure a source to not use a proxy.