(GH-1165) Config - Proxy Bypass List
Add a proxy bypass list to the configuration so that proxies are able
to be bypassed for some connections. Without this an explicit proxy
that is set will never bypass settings for downloads.

Ensure proxy bypass is used with NuGet.Core connections to download
packages and when downloading resources in PowerShell scripts.