(GH-1167) Remove quotes before Process call
In CommandExecutor and Start-ChocolateyProcessAsAdmin, sometimes those
can get called with an executable to run that is surrounded by quotes.
Always remove those quotes so that the process will not fail with
messages similar to:

`Exception calling "Start" with "0" argument(s): "The parameter is
incorrect"`

Without this change, a user needs to ensure that the quotes are
stripped out manually, which can sometimes be hard to determine that is
what needs to be done based on the cryptic error message.