(#1176) Extend Install-ChocolateyFileAssociation
This commit adds additional parameter to the function, with the default
value that maintains full backwards compatibility.

This allows for greater control over the arguments that are passed
through when setting up the file association.  This allows for scenarios
where you might need to do something like `-o "%1"` when the default is
to do something like `"%1" "%*"`.

Co-authored-by: bazzilic <bazzilic@gmail.com>