(GH-1193) Allow version check with FIPS
There are some configuration management tools that need to check the
version of Chocolatey prior to allowing it to set features. This
presents an issue when attempting to set FIPS compliance, because any
call to choco that isn't enabling the FIPS compliance checksum feature
will fail. This should allow a call to find out the version or the
default help without error.