(build) Fix: mono builds
Due to a recent change to force the nant framework, TravisCI started
failing on the following:

`Target framework could not be changed. "net-4.0" is not a valid
framework identifier. Valid values are: mono-2.0, moonlight-2.0,
mono-1.0, mono-4.0, mono-3.5.`

Fix the build so that it overrides the framework when run through mono.