(GH-1054) Externalize the app.manifest
If an app.manifest is embedded in an assembly, you can not override it
at all with a local manifest created next to it. However there is one
scenario that does allow for it to be adjusted, set the file outside
next to the executable to control the behavior of the assembly.

This allows it to be adjusted when folks need to control the behavior
of Chocolatey to suit their needs. This is a compromise to ensure good
default behavior with the ability to adjust it when necessary.