(GH-1203) Ensure PowerShell uses CRLF
Ensure all PowerShell scripts are CRLF and UTF-8 so that running
Chocolatey with an "AllSigned" execution policy does not result in any
errors. There is a bug in PowerShell that doesn't allow it to verify
scripts that have LF file endings. It adds the block of the Authenticode
signature with CRLF line endings every time, resulting in a mixed EOL
format.

When signing scripts that end in LF, PowerShell will see the signature
as valid, but it will fail running the script with 'filename "may have
been tampered because the hash of the file does not match the hash
stored in the digital signature."' However the file had not been
changed. This has been determined to be a bug in PowerShell and has
been shown to affect v3 to v5, but likely also to affect v2 and the
upcoming v6.