(GH-797) Encrypt & remember certain arguments
Ensure the entire remembered argument set is encrypted. There may be
some information that was passed through that that may be sensitive
that is not easily seen by arguments that are known as sensitive.

Fix up the arguments being used to ensure that only a limited set are
being remembered to start. Things that a user would definitely want to
be remembered. Leave out some that were specific to the install, like
version or whether checksums were ignored.