(GH-1210) CacheLocation - Prevent nested chocolatey directories
This ensures that the CacheLocation only appends chocolatey to the TEMP variable
if it has not already been added. Otherwise multiple package installs in the
same process result in an ever deeper nested chocolatey cache folders, resulting
in eventual `PATH too long` errors. This appears to affect only API usage of
Chocolatey.