(GH-1212) Fix: downgrade should not run w/out version
When adding in `--allow-downgrade` to your upgrade options and you have
a prerelease installed, it should not attempt to downgrade to the last
stable version, unless you explicitly pass the version you want it to
upgrade (or downgrade in this case) to. When version is null, it means
that it should function as if it is on the latest release and do
nothing.