(GH-1225) Ensure scripts are UTF-8 BOM
PowerShell is not very good at telling the difference between ANSI
and UTF8 w/out BOM, so PowerShell scripts are not able to be verified
by what is built into PowerShell authenticode verification if they use
UTF8 no BOM with unicode characters. This follows up on #1209 where it
was discovered that the authenticode verification would not work with
LF line endings. That was resolved and later a unicode character
(copyright) was introduced, which broke the functionality again.

Convert all scripts to UTF8 w/BOM.