FIX: Install-ChocolateyPath won't add a path that is a substring of an existing path
New Feature: Check if path exists on disk before adding it.  Default behavior is to ignore the install if the path doesnt exist.  A Force parameter was added for weird cases where that would be desirable.

New Feature: Install-ChocolateyPath now understands cmd-type variables.

A complete pester test file is available.  I didnt add it to the commit as I haven't seen any for powershell methogs in the project.