(GH-1294) API - Expose container directly
Sometimes you need to get to an instance of something provided by
Chocolatey that is loaded into the container. Provide a way to get to
instances of items loaded on the container by accessing the container
directly.

Provide warning that once the container is accessed, no more items can
be registered on the container.