(GH-1296) API - Reuse existing base configuration
Previously, every time the configuration was called for in the API it
would return a new `ChocolateyConfiguration` object. However this
would not work with licensed editions of Chocolatey, as they load some
base information into that configuration and need to ensure that they
can load that base configuration each time and licensed functionality.

Ensure instead that the configuration gets reset to the base each time
it is used and doesn't lose that base information.