(GH-389) API - Ensure ChocolateyInstall env var
When running Chocolatey, it is helpful to ensure environment variables
are set correctly. Sometimes ChocolateyInstall location will not have
been determined correctly or be waiting on a reboot/explorer.exe
restart prior to applications realizing that the value has changed.
Ensure that even with that, the value for ChocolateyInstall is
determined correctly.