(maint) API - Warn on extraction errors
If extraction fails for internal resources, the API should log an Error
but not fail the process. Other parts of the application could fail the
process later and that is expected if things are missing. However
extraction could occur without proper permissions and the API should
just log that it was unable to as not every call through Chocolatey is
going to need those resources.