(GH-1285) Default push to new or old w/warning
When pushing larger packages, the original push url was subject to a
timeout of 100 seconds, a limitation of using Cloudflare. The error:
`Error publishing package. NuGet server returned 524: Origin Time-out"
when pushing through the Cloudflare infrastructure. Instead, convert to
a new default push url that is not subject to a timeout by Cloudflare.

To ensure this is not a breaking change, determine whether the old
apikey was set and use it with a warning to change it over to use the
new url.