(GH-1304) Wrapping HashTable calls inside parenthesis
Previously when running choco new the generated uninstall script would
create a elseif statement which would output several warnings to the
user if more than the expected registry keys was found.
These warnings contains an incorrect use when calling properties
on a hashtable and would instead output the full registry key and their
subkeys instead of the number of keys found, and their display name.

With this change the uninstall template will correctly generate the
use of such properties and output the actual number of keys found
and their display name.

fixes #1304