(GH-1241) More Robust File.Replace
Using File.Replace can have some access violation concerns. It demands
full access to a file, which can cause issues if the file is locked. If
a file is open in another process, it has issues renaming the file,
which causes the rest to not work appropriately.

Implement file replace methods in a move step that will work with files
that are open and then copy the original file to the destination and
delete the original file.

If the file locked is the backup file, it should log that it was not
able to produce the backup and continue without failure. If it has
issues removing the source file, it should log that it was not able
to delete the file and continue without failure.