(GH-1292) Ensure manifest extracted on unpackself
If the `choco.exe.manifest` file is not present on first use of
choco.exe, it extracts the file. However Windows will have already
cached that choco.exe doesn't have a manifest and will not recheck
until the choco.exe changes (last modified time) as it caches that it
doesn't have a manifest. To force a recheck, someone would need to
"touch" the file to change the last modified date.

The issue comes down to the fact that a manifest file needs to be in
place when choco is installed, before the first time that `choco.exe`
gets called. Ensure that the manifest file is packaged up and it will
be unpacked automatically next to choco at the right time.