(GH-1315) Skip AutoUninstaller if no uninstall string
When there is no uninstall string, auto uninstaller will not be able to
do anything with uninstalling the application. Instead it should note
that it is skipping the uninstall of an application (with the display
name) and continue on.

Also note the display name if available for auto uninstaller
information.