(GH-1248) Fix - multiple sources w/same base url fail
Previously, when matching to an authenticated source that uses
credentials, Chocolatey used the first source it found, even if there
may have been multiple sources. Instead build a list of candidate
sources by matching host (base url), and if multiple sources are found,
then match against the full source url for a match. If that
fails, fall back to the previous behavior of using the first found
source.