(GH-1326) Always show heading
When the released version is four digits long, you don't see the heading.
This is due to a change for 0.9.9.9 (59dba8c) that removed the heading.
However, when 0.9.10 came out over a year ago, the heading started showing
up again. When a version came out with a patch recently that was four digits
long, the heading disappeared again. Show the heading always unless limiting
output.