(GH-1325) List/Search - no fail on local directory
When looking at listing verbose information as you see with `choco
info`, if the directory was a local file directory, the command would
fail with the following error:
`System.UriFormatException: Invalid URI: The format of the URI could not
be determined.`

This is due to the way that choco loads up package results, attempting
to convert sources to Uris. Local file and relative directories don't
adjust to a URI very well and cause the above error. When that happens,
simply use the source as it is already set.