(GH-1323) File/File64 as params not aliases
File and File64 were added as aliases to `Url`/`Url64bit` in 0.10.6 to
add compatibility, but they actually introduced more incompatibilities
instead of addressing them due to being aliases. Move them to
parameters and set the values if the original values are empty.

NOTE: This could still cause a small incompatibility if someone is
attempting to use both File and Url in a single set of package
parameters but apply it to multiple function calls. Fortunately it
typically means they will have filled out both items, which means those
functions should still continue to work as they have in the past.