(GH-1328) Serialize/deserialize package files warnings
Instead of defaulting to errorrs, silently have those warnings come up.
When installing the package, if the save of the files snapshot fails,
log a warning but allow the install to be successful.

This corrects seeing these kinds of errors:

~~~
Error deserializing response of type
chocolatey.infrastructure.app.domain.PackageFiles:
 '.', hexadecimal value 0x00, is an invalid character. Line 1, position
1.
~~~